moved language and theme settings to the appearance tab
