+++ src/modals/AppSettingsModal/AppSettingsModal.General.js
@@ -1,13 +1,10 @@
 import React, {
-  memo, useState, useEffect, useMemo,
+  memo, useState, useEffect,
 } from 'react'
 import { useDispatch, useSelector } from 'react-redux'
 import { Checkbox } from '@ufx-ui/core'
 import { useTranslation } from 'react-i18next'
-import _values from 'lodash/values'
-import _map from 'lodash/map'
 
-import Dropdown from '../../ui/Dropdown'
 import WSActions from '../../redux/actions/ws'
 import GAActions from '../../redux/actions/google_analytics'
 import UIActions from '../../redux/actions/ui'
@@ -17,7 +14,7 @@ import {
   updateAutoLoginState,
 } from '../../util/autologin'
 import {
-  SETTINGS, getDMSSetting, getGASetting, getShowAlgoPauseInfoSetting, getRebootSetting, getThemeSetting, THEMES,
+  SETTINGS, getDMSSetting, getGASetting, getShowAlgoPauseInfoSetting, getRebootSetting,
 } from '../../redux/selectors/ui'
 import { DONT_SHOW_DMS_MODAL_KEY } from '../../constants/variables'
 
@@ -28,7 +25,6 @@ const General = () => {
   const { t } = useTranslation()
   const settingsDms = useSelector(getDMSSetting)
   const settingsGa = useSelector(getGASetting)
-  const settingsTheme = useSelector(getThemeSetting)
   const settingsShowAlgoPauseInfo = useSelector(getShowAlgoPauseInfoSetting)
   const settingsRebootAutomatically = useSelector(getRebootSetting)
 
@@ -37,12 +33,6 @@ const General = () => {
   const [isGaChecked, setIsGaChecked] = useState(settingsGa)
   const [isRebootChecked, setIsRebootChecked] = useState(settingsRebootAutomatically)
   const [isShowAlgoPauseInfoChecked, setIsShowAlgoPauseInfoChecked] = useState(settingsShowAlgoPauseInfo)
-  const [currentTheme, setCurrentTheme] = useState(settingsTheme)
-
-  const themes = useMemo(() => _map(_values(THEMES), (value) => ({
-    label: t(`appSettings.${value}`),
-    value,
-  })), [t])
 
   useEffect(() => {
     setIsDmsChecked(settingsDms)
@@ -51,10 +41,6 @@ const General = () => {
   useEffect(() => {
     setIsGaChecked(settingsGa)
   }, [settingsGa])
-
-  useEffect(() => {
-    setCurrentTheme(settingsTheme)
-  }, [settingsTheme])
 
   useEffect(() => {
     setIsShowAlgoPauseInfoChecked(settingsShowAlgoPauseInfo)
@@ -76,13 +62,6 @@ const General = () => {
     setIsGaChecked(nextGa)
     dispatch(WSActions.saveSettings(SETTINGS.GA, nextGa))
     dispatch(GAActions.updateSettings())
-  }
-
-  const updateTheme = (nextTheme) => {
-    setCurrentTheme(nextTheme)
-    dispatch(WSActions.saveSettings(SETTINGS.THEME, nextTheme))
-    dispatch(GAActions.updateSettings())
-    localStorage.setItem(SETTINGS.THEME, nextTheme)
   }
 
   const updateAOPause = (nextAOPause) => {
@@ -127,14 +106,6 @@ const General = () => {
           label={t('appSettings.usageReportingCheckbox')}
           checked={isGaChecked}
           className='appsettings-modal__checkbox'
-        />
-      </div>
-      <div className='appsettings-modal__setting appsettings-modal__dropdown'>
-        <Dropdown
-          label={t('appSettings.themeSetting')}
-          onChange={updateTheme}
-          value={currentTheme}
-          options={themes}
         />
       </div>
       <div className='appsettings-modal__setting'>
